* configure.ac : Fix typo in use_gtk2_4.  Thanks to Mamoru   KOMACHI.
* uim/context.h : Fix bug #3252 in correct handling of return
  vaule of bind_textdomain_codeset().

* xim/xim.h : Remove unused parameters in definition of
  Connection::xim_disconnect(), Connection::xim_sync_reply(),
  icxatr::set_atr().
* xim/ximtrans.cpp : Remove unused parameter in
  Connection::xim_disconnect(), Connection::xim_sync_reply().
* xim/ximic (icxatr::set_atr) : Remove unused parameter.
  Terminate string appropriately.
* xim/ximserver.cpp (print_ustring) : Fix compiler warning.
* xim/main.cpp (X_ErrorHandler) : Show detailed error message.
(X_IOErrorHandler) : Ditto.
(init_supported_locales) : Fix indentation.
* xim/convdisp.cpp (create_default_fontset) : Remove unused
  parameter.
(choose_default_fontset) : Ditto.
(PeWin::expose) : Fix compiler warning.

+++ xim/ximserver.cpp
@@ -75,7 +75,7 @@ static int gXNumLockMask;
 void print_ustring(uString *s)
 {
     uString::iterator i;
-    printf("length=%d : ", s->size());
+    printf("length=%d : ", (int)s->size());
     uchar ch;
     char utf8[6];
     int nbyte;
